initial Primer::LinkComponent to Primer::Beta::LinkComponent migration
