+++ lib/tasks/docs.rake
@@ -64,7 +64,7 @@ namespace :docs do
       Primer::Alpha::HiddenTextExpander,
       Primer::LabelComponent,
       Primer::LayoutComponent,
-      Primer::LinkComponent,
+      Primer::Beta::Link,
       Primer::Markdown,
       Primer::MenuComponent,
       Primer::Navigation::TabComponent,
@@ -110,7 +110,7 @@ namespace :docs do
       Primer::Alpha::Tooltip,
       Primer::ButtonComponent,
       Primer::IconButton,
-      Primer::LinkComponent,
+      Primer::Beta::Link,
       Primer::Alpha::ToggleSwitch,
       Primer::Alpha::ActionList,
       Primer::Alpha::NavList,
