+++ test/components/component_test.rb
@@ -82,7 +82,7 @@ class PrimerComponentTest < Minitest::Test
     [Primer::Alpha::HiddenTextExpander, { "aria-label": "No action" }],
     [Primer::LabelComponent, {}],
     [Primer::LayoutComponent, {}],
-    [Primer::LinkComponent, { href: "https://www.google.com" }],
+    [Primer::Beta::Link, { href: "https://www.google.com" }],
     [Primer::Markdown, {}],
     [Primer::MenuComponent, {}, proc { |c| c.item(href: "#url") { "Item" } }],
     [Primer::Navigation::TabComponent, {}],
@@ -109,6 +109,7 @@ class PrimerComponentTest < Minitest::Test
 
   def test_registered_components
     ignored_components = [
+      "Primer::LinkComponent",
       "Primer::ImageCrop",
       "Primer::Image",
       "Primer::Alpha::ActionList::Heading",
